requirements.txt
Django==4.0.2
pymongo==3.12.1
djongo
pytz
dnspython
python-dotenv
django-grappelli
django-cors-headers
django-tinymce
web3
sendgrid